3859 добавил логику во вьюху и вью модель ручного создания платежа - добавил признак ручного платежа; - добавил последними столбцами инфу о ручном платеже в журнал и отчет по оплатам (ФО); - добавил в фильтр журнала платежей возможность выборки ручных платежей; - перенес из библиотеки словарь для определения типа клиента; - перенес вью модели ручного распределения и загрузки выписки в проект с вью моделями;
