+++ Vodovoz/ServiceDialogs/LoadFrom1cDlg.cs
@@ -6,8 +6,8 @@ using System.Xml;
 using Gamma.GtkWidgets;
 using Gtk;
 using QS.Banks.Domain;
+using QS.Banks.Repositories;
 using QS.DomainModel.UoW;
-using QSBanks.Repositories;
 using QS.BusinessCommon.Domain;
 using QS.BusinessCommon.Repository;
 using QSProjectsLib;
