+++ VodovozInfrastructure/InformationHandbook.cs
@@ -1,0 +1,43 @@
+﻿using System.Collections.Generic;
+
+namespace VodovozInfrastructure
+{
+	public class InformationHandbook
+	{
+		public static Dictionary<string, string> OrganizationTypes = new Dictionary<string, string> {
+			{ "ООО", "Общество с ограниченной ответственностью" },
+			{ "ЗАО", "Закрытое акционерное общество" },
+			{ "ОАО", "Открытое акционерное общество" },
+			{ "ГорПО", "Городское потребительское общество" },
+			{ "РайПО", "Районное потребительское общество" },
+			{ "СельПО", "Сельское потребительское общество" },
+			{ "НКО", "Некоммерческая организация" },
+			{ "ИП", "Индивидуальный предприниматель" },
+			{ "ОДО", "Общество с дополнительной ответственностью" },
+			{ "ПТ", "Полное товарищество" },
+			{ "АО", "Акционерное общество" },
+			{ "АК", "Акционерная компания" },
+			{ "КТ", "Коммандитное товарищество" },
+			{ "ХП", "Хозяйственное партнерство" },
+			{ "КФХ", "Крестьянское (фермерское) хозяйство" },
+			{ "НП", "Некоммерческое партнерство" },
+			{ "ПК", "Потребительский кооператив" },
+			{ "ТСН", "Товарищество собственников недвижимости" },
+			{ "АНО", "Автономная некоммерческая организация" },
+			{ "ФГУП", "Федеральное государственное унитарное предприятие" },
+			{ "ГУП", "Государственное унитарное предприятие" },
+			{ "МУП", "Муниципальное унитарное предприятие" },
+			{ "ГП", "Государственное предприятие" },
+			{ "СНТ", "Садоводческое некоммерческое товарищество" },
+			{ "ДНТ", "Дачное некоммерческое товарищество" },
+			{ "ОНТ", "Огородническое некоммерческое товарищество" },
+			{ "СНП", "Садоводческое некоммерческое партнерство" },
+			{ "ДНП", "Дачное некоммерческое партнерство" },
+			{ "ОНП", "Огородническое некоммерческое партнерство" },
+			{ "СПК", "Садоводческий потребительский кооператив" },
+			{ "ДПК", "Дачный потребительский кооператив" },
+			{ "ОПК", "Огороднический потребительский кооператив" }
+		};
+	}
+}
+
